add widget filter
refactor pick widget activity code
fix widget picker list layout
