+++ app/src/main/java/rocks/tbog/tblauncher/widgets/MenuItem.java
@@ -1,0 +1,8 @@
+package rocks.tbog.tblauncher.widgets;
+
+import androidx.annotation.NonNull;
+
+interface MenuItem {
+    @NonNull
+    String getName();
+}
